+++ src/game/ui/menu.tsx
@@ -88,7 +88,7 @@ export function Menu(props: MenuProps) {
         setPage(page.kind === 'changelog' ? { kind: 'home' } : { kind: 'changelog' })
       }}
     >
-      2.4.0
+      2.4.1
     </button>
   )
   return (